requirements.txt
numpy==1.18.1
pygame==1.9.6
PyGLM==1.1.7
PyOpenGL==3.1.4
PyOpenGL-accelerate==3.1.4